move everything stream related to it's own folder
