+++ src/Sound/Tidal/Context.hs
@@ -22,7 +22,7 @@ import Prelude hiding ((<*), (*>))
 
 import Data.Ratio as C
 
-import Sound.Tidal.Config as C
+import Sound.Tidal.Stream.Config as C
 import Sound.Tidal.Control as C
 import Sound.Tidal.Core as C
 import Sound.Tidal.Params as C
@@ -31,7 +31,7 @@ import Sound.Tidal.Pattern as C
 import Sound.Tidal.Scales as C
 import Sound.Tidal.Show as C
 import Sound.Tidal.Simple as C
-import Sound.Tidal.Stream as C
+import Sound.Tidal.Stream.Stream as C
 import Sound.Tidal.Transition as C
 import Sound.Tidal.UI as C
 import Sound.Tidal.Version as C
